Create V2 Model and Schema
Add tags for permits and regulations using linking tables. Tags will be used for search on the website.
Add regulationNotes and permitNotes columns that store user generated TEXT for permits and regulations. Requested due to the automated nature of web scrapers, and potential to lose information added by DB administrators.
Renamed the Permits table renewalPeriod to validFor for clarification.